Update use-color for native
