--- packages/block-editor/src/components/colors/use-colors.native.js
@@ -1,10 +1,0 @@
-const TextColor = ( props ) => <>{ props.children }</>;
-
-const InspectorControlsColorPanel = () => null;
-
-export default function __experimentalUseColors() {
-	return {
-		TextColor,
-		InspectorControlsColorPanel,
-	};
-}
